requirements.txt
Flask==2.3.3
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3
matplotlib==3.10.7
seaborn==0.13.2
tqdm==4.67.1
loguru==0.7.3
python-dotenv==1.1.1
Pillow==10.0.0
requests

# PyTorch + CUDA if available, otherwise CPU
torch
torchvision
torchaudio
